Add files via upload
Design and Construction of Numerical Overcurrent Relay

The project aims at the design and the construction of a laboratory replica for a commercial
numerical over current relay using a microcontroller and micro CT. It will acquire, sample and
convert into digital form the analog current signals from a power system in real time. Then it
simulates the working principles of a conventional electromechanical overcurrent relay in
through a numerical (mathematical) model. Appropriate equations and algorithm have been used.
When the current in a part of a power system exceeds a pre-set value either due to a fault or a large
overload the relay will isolate the part through sending a trip signal to a circuit breaker. In the
developed laboratory model, the system is assembled by a lamp board and the circuit breaker is
simulated by a hardware solid state (MOSFET) switch. The constructed relay has been tested
extensively for various values of load current and two different tripping characteristics (Inverse
Definite Minimum Time: IDMT) have been verified.